Separate the commonly used schema functions into separate file #32
Fixed a couple of  faulty functions on the way. Need ideas for more
utilities!
